requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
Werkzeug==2.3.7
deepface==0.0.79
tensorflow==2.13.0
opencv-python==4.8.1.78